Add test_xform2json file
- Move 'test_convert_toJSON_multi_language' test
